+++ simulator/examples/offscreen.rs
@@ -1,7 +1,7 @@
 use embedded_graphics::{
     pixelcolor::BinaryColor, prelude::*, primitives::Rectangle, style::PrimitiveStyle,
 };
-use embedded_graphics_simulator::{SimulatorDisplay, WindowBuilder};
+use embedded_graphics_simulator::{OutputSettingsBuilder, SimulatorDisplay, Window};
 
 fn main() -> Result<(), core::convert::Infallible> {
     let mut display: SimulatorDisplay<BinaryColor> = SimulatorDisplay::new(Size::new(32, 32));
@@ -12,12 +12,8 @@ fn main() -> Result<(), core::convert::Infallible> {
         .translate(Point::new(-8, -8))
         .draw(&mut display)?;
 
-    let mut window = WindowBuilder::new(&display)
-        .title("Offscreen")
-        .scale(4)
-        .build();
-
-    window.show_static(&display);
+    let output_settings = OutputSettingsBuilder::new().scale(4).build();
+    Window::new("Offscreen", &output_settings).show_static(&display);
 
     Ok(())
 }